ADD class rt_Class to replace struct rt_Class
To keep header file clean and neat, c++ is more and more needed to
handle the situation. ClassFile will be, after its loading, linked
and initialized, as if it is treated in a real JVM. In this procedure,
more CPU time and more memory space are assumed to be cosumed.